+++ CHANGELOG.md
@@ -118,6 +118,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Desert area wildlife is back to spawning in hot temperatures.
 - Ambience (e.g. wind) loops properly.
 - Can no longer get bags with items still in them into the inventory by "using" another bag.
+- Npcs don't get angry if you kill a monster infront of them, and can get happy if you saved them from it.
 
 ## [0.17.0] - 2024-12-28